+++ R/funcharts.R
@@ -19,10 +19,11 @@
 #' @importFrom stats
 #' model.matrix rstandard formula uniroot
 #' @importFrom tidyr pivot_longer drop_na
-#' @importFrom rlang .data
+#' @importFrom rlang .data :=
 #' @importFrom RSpectra eigs_sym
 #' @importFrom matrixStats rowCumsums
 #' @importFrom graphics par persp
+#' @importFrom stringr str_count
 #' @references
 #' Capezza C, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2020)
 #' Control charts for